dont recursively call find_bewteen
